First attempt at gait recognition when walking toward the Kinect sensor.
The xlsx files are gait signatures for my wife and I after walking in front of a Kinect sensor 10 times. The X axis represents the different joints (head, arms, legs) and the Y axis is the distance between the joints. Naturally there is more noise on the left and right legs because that is what moves the most while you are walking.

Reviewed by: fatjawa